extract feature tests to separate module
